Two more enemies added
Created an enemy class which is the parent one to all enemies. Created 2 more enemies, one which tries to reach the previous position of the main character (to try to get behind it in case it tries to escape from the first enemy) and another one that moves randomly between the level exterior corners. A functionality was added to all enemies for them to run to a corner when the main character picks the power up (the power up is activated using the spacebar, for now).

A bug in the adjacency matrix which caused the characters to cross walls and get out of the screen was fixed .