Dying and power ups added
Dying possibilities for the characters and power pellets were added. Nevertheless, the powerup is infinite and the enemies do not change their vulnerable state when they die. The powerup can be manually deactivated using spacebar. The enemies respawn on fixed positions of the level when they die, this will be changed in the future.